Update Devise
Devise was blocking the app from booting if the Devise key wasn't set.
We want to get that from the credentials.yml.enc file, but if you're
setting up the app for the first time you don't have this and it creates
a circular dependency.

The updated version of Devise makes it so that you can specify a
temporary secret key in the shell environment and let Rails generate
baseline credentials for you.